Rename callapi/call_api_csv.py to bin/callapi.py
These names are more in the spirit of what these scripts are
intended to do.  Because the name no longer suggests CSV output
specifically, I will add functionality in the future to allow the
user to be able to choose the output format.  For now, however, the
default output format is still CSV.