Switch from command line args to a configuration file.
In its current state, the program looks for a file path in the
CANVASLMS_API_CONFIG_PATH environment variable.  If found, it reads
in the JSON-encoded configuration map object.

The three specific configuration values that the program looks for at
this point are (case-sensitive):
 - TokenPath: the path of the file containing the Canvas API
   authentication token.
 - CavnasHost: the name of the host running the instance of Canvas for
   which you'd like to access the API.
 - APICall: the actual call to make (e.g., "accounts/123/courses")

Currently the only way to specify an API call is to set the APICall
value in the configuration file.  This will change shortly.