Add support for configuration files and command line args.
Something that made this script really klunky to use in the past
was that, in addition to the actual API call you were trying to make,
you also had to specify the Canvas host and the path to the
authentication token file on the command line.

By adding functionality for configuration files and for command
line options, I've made the process more efficient and user-friendly.

For example, if a proper configuration file is used one can make a
call to the API with an invocation as simple as follows:

  python3 -m canvaslms.bin.callapi account/123456/courses