Add function to provide auth token on the CLI.
For the use case where the user doesn't store (or _want to_ store)
the authentication token in a file.  For example, the user may wish to
store the auth token in an encrypted password database (e.g., KeePass)
and copy and paste the auth token only when necessary.

To enable command line prompt fro the auth token, use the -A option.

Currently, even when using the -A option, the -T option (specify a
token file) is still mandatory (although the file isn'at actually used.)
This is a bug and will be changed in a future release.